removed recreating activity, added replacing webview content - that saved activities path
