+++ app/src/main/java/pl/zhp/natropie/ui/WebView/Reader.kt
@@ -16,6 +16,7 @@ import pl.zhp.natropie.ReaderActivity
 import pl.zhp.natropie.db.NaTropieDB
 import pl.zhp.natropie.db.entities.Post
 import pl.zhp.natropie.db.entities.PostWithColor
+import pl.zhp.natropie.helpers.NaTropiePage
 import pl.zhp.natropie.tracking.Track
 import pl.zhp.natropie.ui.models.PostVM
 
@@ -30,18 +31,17 @@ class Reader(val context: Context, private val activity: AppCompatActivity) : We
 
     override fun shouldOverrideUrlLoading(view: WebView, url: String): Boolean {
         if (!url.startsWith(READER_PREFIX)) {
-            Toast.makeText(context,url, Toast.LENGTH_SHORT).show()
+            Toast.makeText(context, url, Toast.LENGTH_SHORT).show()
             val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(url))
             browserIntent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
             context.startActivity(browserIntent)
             return true
         }
-        reloadActivity(activity, url)
+        reloadActivity(view, url)
         return true
     }
 
-    private fun reloadActivity(activity: AppCompatActivity, url: String) {
-        activity.finish()
+    private fun reloadActivity(view: WebView, url: String) {
         val slug: String = getSlugFromUrl(url)
 
         scope.launch {
@@ -50,16 +50,24 @@ class Reader(val context: Context, private val activity: AppCompatActivity) : We
                 table.getBySlug(slug)
             }
             val post: PostWithColor? = job.await()
-            if (post!==null) {
-                val intent = Intent(activity, ReaderActivity::class.java)
-                intent.putExtra(ReaderActivity.VAR_POST, Parcels.wrap(post))
-                intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-                activity.startActivity(intent)
+            if (post !== null) {
+                loadPage(view, post)
             } else {
-                Toast.makeText(context,"Nie znaleziono artykułu",Toast.LENGTH_LONG).show()
+                Toast.makeText(context, "Nie znaleziono artykułu", Toast.LENGTH_LONG).show()
             }
         }
+    }
 
+    fun loadPage(webView: WebView, post: PostWithColor) {
+        val doc = NaTropiePage(post.content)
+        doc.setTitle(post.title)
+        if (post.category != "PAGES") {
+            doc.setAuthor(post.author)
+                .setCategory(post.category)
+                .setDate(post.date)
+        }
+        Track.DisplayPost(post.id, post.title, post.author, post.category)
+        webView.loadDataWithBaseURL("file:///android_asset/", doc.getHtml(), "text/html", "UTF-8", "")
     }
 
     private fun getSlugFromUrl(url: String): String {
